feat: call the model by api
