+++ config.py
@@ -2,6 +2,7 @@ api_map = {
     "ModelSelection": "http://127.0.0.1:8000/model_selection",
     "GPT": "http://127.0.0.1:8001/gpt",
     "Diffusion": "http://127.0.0.1:8002/diffusion",
-    "ModelVerification": "http://127.0.0.1:8003/model_verification"
+    "ModelVerification": "http://127.0.0.1:8003/model_verification",
+    "TextModel": "http://127.0.0.1:8004/text_model"
 }
